Rename benchmark files
